Deleting the admin user is not allowed

+++ soul-admin/src/main/java/org/dromara/soul/admin/service/impl/DashboardUserServiceImpl.java
@@ -116,6 +116,9 @@ public class DashboardUserServiceImpl implements DashboardUserService {
     public int delete(final List<String> ids) {
         int dashboardUserCount = 0;
         for (String id : ids) {
+            if (dashboardUserMapper.selectById(id).getUserName().equals("admin")) {
+                continue;
+            }
             dashboardUserCount += dashboardUserMapper.delete(id);
         }
         return dashboardUserCount;
